requirements.txt
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
greenlet==2.0.2
idna==3.4
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
psycopg2==2.9.5
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
vk-api==11.9.9
